+++ packages/af-utils-react-virtual-headless/src/hooks/useIsomorphicLayoutEffect/index.ts
@@ -1,0 +1,3 @@
+import { useLayoutEffect, useEffect } from "react";
+
+export default process.env.__IS_SERVER__ ? useEffect : useLayoutEffect;