+++ packages/af-utils-react-virtual-headless/src/hooks/useScroller/index.ts
@@ -1,0 +1,10 @@
+import useIsomorphicLayoutEffect from "hooks/useIsomorphicLayoutEffect";
+import type { PublicList, ScrollElement } from "models/List";
+
+const useScroller = (model: PublicList, scroller: ScrollElement) =>
+    useIsomorphicLayoutEffect(() => {
+        model.setScroller(scroller);
+        return () => model.setScroller(null);
+    }, [model, scroller]);
+
+export default useScroller;