+++ packages/website/components/examples/hook/Simple/index.js
@@ -18,7 +18,7 @@ const SimpleHook = () => {
     });
 
     return (
-        <div className="h-full overflow-auto" ref={model.setScrollElement}>
+        <div className="h-full overflow-auto" ref={model.setScroller}>
             <Subscription model={model}>
                 {() => {
                     const fromOffset = model.getOffset(model.from);