+++ packages/af-utils-react-mobx-table/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @af-utils/react-mobx-table
 
+## 1.0.0
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.1.0
+    -   @af-utils/react-table@1.0.0
+
 ## 0.0.58
 
 ### Patch Changes