+++ packages/af-utils-react-virtual-list/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @af-utils/react-virtual-list
 
+## 1.0.0
+
+### Minor Changes
+
+-   upgraded headless lib to typescript, added window scroll
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @af-utils/react-virtual-headless@0.1.0
+
 ## 0.0.52
 
 ### Patch Changes